visual tests improvements
