--- scripts/visual-tests-upload.mjs
@@ -1,16 +1,0 @@
-/* eslint-disable max-len */
-import { spawnProcess } from './utils/processes.mjs';
-
-const commitSha = (await spawnProcess('git rev-parse HEAD', { silent: true }))
-  .stdout
-  .toString()
-  .slice(0, 7);
-
-console.log('Upload to Argos');
-await spawnProcess('npx @argos-ci/cli upload tests/screenshots');
-
-console.log('Upload to Viswiz');
-await spawnProcess(`npx viswiz build --image-dir ./tests/screenshots --message last-commit-message --revision ${commitSha}`);
-
-console.log('Upload to Percy');
-await spawnProcess('npx percy upload ./tests/screenshots');
