+++ visual-test/package.json
@@ -4,8 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "run": "node ../scripts/visual-tests-build.mjs",
-    "upload": "node ../scripts/visual-tests-upload.mjs"
+    "test": "node ./scripts/visual-tests-build.mjs",
+    "upload": "node ./scripts/visual-tests-upload.mjs"
   },
   "keywords": [],
   "author": "",
@@ -14,6 +14,7 @@
     "@argos-ci/cli": "^0.3.1",
     "@percy/cli": "^1.16.0",
     "@playwright/test": "^1.27.1",
-    "viswiz": "^2.2.2"
+    "viswiz": "^2.2.2",
+    "execa": "^5.0.0"
   }
 }
