+++ visual-test/scripts/visual-tests-build.mjs
@@ -1,0 +1,11 @@
+// import { spawnProcess } from './utils/processes.mjs';
+import execa from 'execa';
+
+const wrappers = ['js', 'angular', 'react', 'vue'];
+const version = '12.2.0';
+
+await execa.command('npx playwright install --with-deps');
+
+for (let i = 0; i < wrappers.length; ++i) {
+  execa.command('npx playwright test', { env: { WRAPPER: wrappers[i], VERSION: version } });
+}
