+++ visual-test/tests/.empty-test-template.js
@@ -4,13 +4,13 @@ const { test, expect } = require('@playwright/test');
 
 /* ======= */
 
-const version = process.env.VERSION;
-const wrapper = process.env.WRAPPER;
+const version = process.env.VERSION || helpers.defaultHOTVersion;
+const wrapper = process.env.WRAPPER || helpers.defaultHOTWrapper;
 
 const testURL = `https://examples.handsontable.com/examples/${version}/docs/${wrapper}/demo/index.html`;
 const expectedPageTitle = /Handsontable for .* example/;
 
-const testTitle = '';
+const testTitle = 'Test this and test that';
 const stylesToAdd = [
   helpers.cssFiles.cookieInfo,
   helpers.cssFiles.dynamicDataFreeze
